partial implementation of planner contacts.
